+++ src/test/java/com/tms/csp/pl/SimplifySeriesModelAndsTest.java
@@ -4,6 +4,8 @@ import com.tms.csp.ast.Csp;
 import com.tms.csp.util.CspBaseTest2;
 import org.junit.Test;
 
+import java.math.BigInteger;
+
 import static junit.framework.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
@@ -19,11 +21,11 @@ public class SimplifySeriesModelAndsTest extends CspBaseTest2 {
         csp1 = csp1.condition("YR_2013 !YR_2014");
 
         int size1 = csp1.serialize().length();
-        long count1 = csp1.toDnnf().getSmooth().getSatCount();
+        BigInteger count1 = csp1.toDnnf().getSmooth().getSatCount();
 
         csp1.simplifySeriesModelAnds();
         int size2 = csp1.serialize().length();
-        long count2 = csp1.toDnnf().getSmooth().getSatCount();
+        BigInteger count2 = csp1.toDnnf().getSmooth().getSatCount();
 
         assertEquals(count1, count2);
         assertTrue(size2 < size1);
